+++ src/ios/AppboyPlugin.m
@@ -1,9 +1,10 @@
 #import "AppboyPlugin.h"
-#import <AppboyKit.h>
-#import <ABKAttributionData.h>
+#import "AppboyKit.h"
+#import "ABKAttributionData.h"
 #import "AppDelegate+Appboy.h"
 #import "IDFADelegate.h"
-#import <AppboyNewsFeed.h>
+#import "AppboyNewsFeed.h"
+#import "AppboyContentCards.h"
 
 @interface AppboyPlugin()
   @property NSString *APIKey;
@@ -70,7 +71,7 @@
         options = options | UNAuthorizationOptionProvisional;
       }
       [center requestAuthorizationWithOptions:options
-                            completionHandler:^(BOOL granted, NSError * _Nullable error) {
+                            completionHandler:^(BOOL granted, NSError *_Nullable error) {
                               NSLog(@"Permission granted.");
                               [[Appboy sharedInstance] pushAuthorizationFromUserNotificationCenter:granted];
                             }];
@@ -298,6 +299,15 @@
   [self.viewController presentViewController:newsFeed animated:YES completion:nil];
 }
 
+- (void) launchContentCards:(CDVInvokedUrlCommand *)command {
+  [[Appboy sharedInstance] requestContentCardsRefresh];
+  ABKContentCardsViewController *contentCardsModal = [[ABKContentCardsViewController alloc] init];
+  UIWindow *keyWindow = [[UIApplication sharedApplication] keyWindow];
+  UIViewController *mainViewController = keyWindow.rootViewController;
+  [mainViewController presentViewController:contentCardsModal animated:YES completion:nil];
+}
+
+/*-------News Feed-------*/
 - (void) getNewsFeed:(CDVInvokedUrlCommand *)command {
   [[Appboy sharedInstance] requestFeedRefresh];
   int categoryMask = [self getCardCategoryMaskWithStringArray:command.arguments];
@@ -319,7 +329,6 @@
   [self sendCordovaSuccessPluginResultWithArray:result andCommand:command];
 }
 
-/*-------News Feed-------*/
 - (void) getCardCountForCategories:(CDVInvokedUrlCommand *)command {
   int categoryMask = [self getCardCategoryMaskWithStringArray:command.arguments];
 
@@ -365,6 +374,135 @@
     }
   }
   return categoryMask;
+}
+
+/*-------Content Cards-------*/
+- (void) requestContentCardsRefresh:(CDVInvokedUrlCommand *)command {
+  [[Appboy sharedInstance] requestContentCardsRefresh];
+}
+
+- (void) logContentCardClicked:(CDVInvokedUrlCommand *)command {
+  NSString *idString = [command argumentAtIndex:0 withDefault:nil];
+  ABKContentCard *cardToClick = [self getContentCardById:idString];
+  if (cardToClick) {
+    [cardToClick logContentCardClicked];
+  }
+}
+
+- (void) logContentCardDismissed:(CDVInvokedUrlCommand *)command {
+  NSString *idString = [command argumentAtIndex:0 withDefault:nil];
+  ABKContentCard *cardToClick = [self getContentCardById:idString];
+  if (cardToClick) {
+    [cardToClick logContentCardDismissed];
+  }
+}
+
+- (void) logContentCardImpression:(CDVInvokedUrlCommand *)command {
+  NSString *idString = [command argumentAtIndex:0 withDefault:nil];
+  ABKContentCard *cardToClick = [self getContentCardById:idString];
+  if (cardToClick) {
+    [cardToClick logContentCardImpression];
+  }
+}
+
+- (void) logContentCardsDisplayed:(CDVInvokedUrlCommand *)command {
+  [[Appboy sharedInstance] logContentCardsDisplayed];
+}
+
+- (void) getContentCardsFromServer:(CDVInvokedUrlCommand *)command {
+  NSNotificationCenter *center = [NSNotificationCenter defaultCenter];
+  NSOperationQueue *mainQueue = [NSOperationQueue mainQueue];
+  [center addObserverForName:ABKContentCardsProcessedNotification object:nil
+                                                   queue:mainQueue usingBlock:^(NSNotification *note) {
+                                                     NSLog(@"Got Content Cards from server callback");
+                                                     BOOL updateIsSuccessful = [note.userInfo[ABKContentCardsProcessedIsSuccessfulKey] boolValue];
+                                                     if (updateIsSuccessful) {
+                                                       [self getContentCardsFromCache:command];
+                                                     }
+                                                   }];
+  [[Appboy sharedInstance] requestContentCardsRefresh];
+}
+
+- (void) getContentCardsFromCache:(CDVInvokedUrlCommand *)command {
+  NSArray<ABKContentCard *> *cards = [[Appboy sharedInstance].contentCardsController getContentCards];
+  
+  NSMutableArray *mappedCards = [NSMutableArray arrayWithCapacity:[cards count]];
+  [cards enumerateObjectsUsingBlock:^(id card, NSUInteger idx, BOOL *stop) {
+     [mappedCards addObject:[AppboyPlugin RCTFormatContentCard:card]];
+  }];
+  
+  [self sendCordovaSuccessPluginResultWithArray:mappedCards andCommand:command];
+}
+
+- (nullable ABKContentCard *)getContentCardById:(NSString *)idString {
+  NSArray<ABKContentCard *> *cards = [[Appboy sharedInstance].contentCardsController getContentCards];
+  NSPredicate *predicate = [NSPredicate predicateWithFormat:@"idString == %@", idString];
+  NSArray *filteredArray = [cards filteredArrayUsingPredicate:predicate];
+  
+  if (filteredArray.count) {
+    return filteredArray[0];
+  }
+  
+  return nil;
+}
+
++ (NSDictionary *) RCTFormatContentCard:(ABKContentCard *)card {
+  NSMutableDictionary *formattedContentCardData = [NSMutableDictionary dictionary];
+  
+  formattedContentCardData[@"id"] = card.idString;
+  formattedContentCardData[@"created"] = @(card.created);
+  formattedContentCardData[@"expiresAt"] = @(card.expiresAt);
+  formattedContentCardData[@"viewed"] = @(card.viewed);
+  formattedContentCardData[@"clicked"] = @(card.clicked);
+  formattedContentCardData[@"pinned"] = @(card.pinned);
+  formattedContentCardData[@"dismissed"] = @(card.dismissed);
+  formattedContentCardData[@"dismissible"] = @(card.dismissible);
+  formattedContentCardData[@"url"] = card.urlString ?: [NSNull null];
+  formattedContentCardData[@"openURLInWebView"] = @(card.openUrlInWebView);
+  
+  formattedContentCardData[@"extras"] = [AppboyPlugin getJsonFromExtras:card.extras];
+  
+  if ([card isKindOfClass:[ABKCaptionedImageContentCard class]]) {
+    ABKCaptionedImageContentCard *captionedCard = (ABKCaptionedImageContentCard *)card;
+    formattedContentCardData[@"image"] = captionedCard.image;
+    formattedContentCardData[@"imageAspectRatio"] = @(captionedCard.imageAspectRatio);
+    formattedContentCardData[@"title"] = captionedCard.title;
+    formattedContentCardData[@"cardDescription"] = captionedCard.cardDescription;
+    formattedContentCardData[@"domain"] = captionedCard.domain ?: [NSNull null];
+    formattedContentCardData[@"type"] = @"Captioned";
+  }
+  
+  if ([card isKindOfClass:[ABKBannerContentCard class]]) {
+    ABKBannerContentCard *bannerCard = (ABKBannerContentCard *)card;
+    formattedContentCardData[@"image"] = bannerCard.image;
+    formattedContentCardData[@"imageAspectRatio"] = @(bannerCard.imageAspectRatio);
+    formattedContentCardData[@"type"] = @"Banner";
+  }
+  
+  if ([card isKindOfClass:[ABKClassicContentCard class]]) {
+    ABKClassicContentCard *classicCard = (ABKClassicContentCard *)card; 
+    formattedContentCardData[@"image"] = classicCard.image ?: [NSNull null];
+    formattedContentCardData[@"title"] = classicCard.title;
+    formattedContentCardData[@"cardDescription"] = classicCard.cardDescription;
+    formattedContentCardData[@"domain"] = classicCard.domain ?: [NSNull null];
+    formattedContentCardData[@"type"] = @"Classic";
+  }
+  
+  return formattedContentCardData;
+}
+
++ (NSString *) getJsonFromExtras:(NSDictionary *)extras {
+  NSError *error;
+  NSData *jsonData = [NSJSONSerialization dataWithJSONObject:extras
+                                                     options:0
+                                                       error:&error];
+  
+  if (!jsonData) {
+    NSLog(@"Got an error in getJsonFromExtras: %@", error);
+    return @"{}";
+  } else {
+    return [[NSString alloc] initWithData:jsonData encoding:NSUTF8StringEncoding];
+  }
 }
 
 - (void) sendCordovaErrorPluginResultWithString:(NSString *)resultMessage andCommand:(CDVInvokedUrlCommand *)command {